backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-multipart
python-dotenv